+++ tests/Feature/DialTest.php
@@ -26,6 +26,69 @@ test('initial call-in default', function () {
     ], false);
 });
 
+test('initial call-in with jft option enabled', function () {
+    $_SESSION['override_jft_option'] = "true";
+    $response = $this->get('/');
+    $response
+        ->assertStatus(200)
+        ->assertHeader("Content-Type", "text/xml; charset=UTF-8")
+        ->assertSeeInOrder([
+            '<?xml version="1.0" encoding="UTF-8"?>',
+            '<Response>',
+            '<Pause length="2"></Pause>',
+            '<Say voice="alice" language="en-US">Test Helpline</Say>',
+            '<Say voice="alice" language="en-US">press one to find someone to talk to</Say>',
+            '<Say voice="alice" language="en-US">press two to search for meetings</Say>',
+            '<Say voice="alice" language="en-US">',
+            'press three to listen to the just for today                    </Say>',
+            '</Gather>',
+            '</Response>'
+        ], false);
+});
+
+test('initial call-in with spad option enabled', function () {
+    $_SESSION['override_spad_option'] = "true";
+    $response = $this->get('/');
+    $response
+        ->assertStatus(200)
+        ->assertHeader("Content-Type", "text/xml; charset=UTF-8")
+        ->assertSeeInOrder([
+            '<?xml version="1.0" encoding="UTF-8"?>',
+            '<Response>',
+            '<Pause length="2"></Pause>',
+            '<Say voice="alice" language="en-US">Test Helpline</Say>',
+            '<Say voice="alice" language="en-US">press one to find someone to talk to</Say>',
+            '<Say voice="alice" language="en-US">press two to search for meetings</Say>',
+            '<Say voice="alice" language="en-US">',
+            'press four to listen to the spiritual principle a day                    </Say>',
+            '</Gather>',
+            '</Response>'
+        ], false);
+});
+
+test('initial call-in with jft and spad option enabled', function () {
+    $_SESSION['override_jft_option'] = "true";
+    $_SESSION['override_spad_option'] = "true";
+    $response = $this->get('/');
+    $response
+        ->assertStatus(200)
+        ->assertHeader("Content-Type", "text/xml; charset=UTF-8")
+        ->assertSeeInOrder([
+            '<?xml version="1.0" encoding="UTF-8"?>',
+            '<Response>',
+            '<Pause length="2"></Pause>',
+            '<Say voice="alice" language="en-US">Test Helpline</Say>',
+            '<Say voice="alice" language="en-US">press one to find someone to talk to</Say>',
+            '<Say voice="alice" language="en-US">press two to search for meetings</Say>',
+            '<Say voice="alice" language="en-US">',
+            'press three to listen to the just for today                    </Say>',
+            '<Say voice="alice" language="en-US">',
+            'press four to listen to the spiritual principle a day                    </Say>',
+            '</Gather>',
+            '</Response>'
+        ], false);
+});
+
 test('initial call-in default with language selections', function () {
     $_SESSION['override_language_selections'] = "en-US,es-US";
     $response = $this->get('/');